docking: Remove DashSpacer and override ControlsManagerLayout logic instead
To compute the workspace box size, we used to add a dummy actor (the
DashSpacer) which was bound to the actual dock size so that it could
have been used by the controls manager layout to figure out the space
used by the dash.

While this may still work in horizontal mode, it does not when using a
vertical dash.

So, let's override workspace box computation code instead, making the
dash size to change depending on its disposition.