Utils: Add class to inject virtual functions
In some cases we need to change the behavior of the upstream virtual
functions, to do this we can't just monkey-patch the definition but we
need to register the new function, so let's handle it.

Given that we don't redefine a replacement for the overridden function,
it's still possible to call the original implementation calling
this.vfunc_overridden_virtual_function in the callback closure function
(not an arrow function!) as we'd wo with `super` in the normal code.