dash: Include paddings and separator to compute available space
To compute the real available space for icons we must take care of the
actual dash container padding, or we'd end up considering more space
than the one that the theme defines. So use the dash container to
compute the allocation box.

Also use the proper child to compute the size difference between the
icon and its parent button, and eventually include the separator (if
set) size to compute the available space.

This fixes the dynamic icon sized mode.