appIcons: Manage app-urgency per windows
An application is urgent when one of its windows is, however we were
resetting this once the application was focused.

While this is acceptable for manual urgency (the one set via APIs for
example), in the case of urgency triggered by a single window we should
keep this state till the window isn't set as non-urgent.

So now:
 - An application keeps stay urgent until any window is requiring
   attention or marked as urgent.
 - An application marked as urgent from unity API is urgent until
   all of its windows have not been focused.