dash: Don't change packing just to reorder dash children (and show Apps)
Changing pack_start in the dash container to order the show-apps button
at top or bottom can be a shortcut, but also causes some unwanted
results, because the items are still kept in the original order and
st-widget won't use the correct first/last-child pseudo classes, leading
to a wrong look.

In fact, right now, when the AppIcon is shown as first icon, there's no
top padding set on it.